+++ lib/src/models/abstract/serializable_model.dart
@@ -1,3 +1,0 @@
-abstract class SerializableModel {
-  Map<String, dynamic> toJson();
-}