+++ lib/src/models/alfred_item_text.dart
@@ -1,17 +1,20 @@
-import 'abstract/serializable_model.dart';
+import 'package:json_annotation/json_annotation.dart';
 
-class AlfredItemText implements SerializableModel {
+part 'alfred_item_text.g.dart';
+
+@JsonSerializable()
+class AlfredItemText {
   const AlfredItemText({
     required this.copy,
     this.largeType,
   });
 
   final String copy;
+  @JsonKey(name: 'largetype')
   final String? largeType;
 
-  @override
-  Map<String, dynamic> toJson() => {
-        'copy': copy,
-        if (largeType != null) 'largetype': largeType,
-      };
+  factory AlfredItemText.fromJson(Map<String, dynamic> json) =>
+      _$AlfredItemTextFromJson(json);
+
+  Map<String, dynamic> toJson() => _$AlfredItemTextToJson(this);
 }