+++ lib/src/models/alfred_item_text.g.dart
@@ -1,0 +1,19 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'alfred_item_text.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+AlfredItemText _$AlfredItemTextFromJson(Map<String, dynamic> json) =>
+    AlfredItemText(
+      copy: json['copy'] as String,
+      largeType: json['largetype'] as String?,
+    );
+
+Map<String, dynamic> _$AlfredItemTextToJson(AlfredItemText instance) =>
+    <String, dynamic>{
+      'copy': instance.copy,
+      'largetype': instance.largeType,
+    };