+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- Use json_serializable to automate json serialization
+- BREAKING: Change Iterable<AlfredItem> for List<AlfredItem>
+
 ## 0.0.1
 
 - Initial version.